Move preloads drop to userData/sentry
