+++ scripts/build-preloads.js
@@ -1,26 +1,28 @@
 const { readFileSync, writeFileSync } = require('fs');
 const { join } = require('path');
-const ts = require('typescript');
+const { transpile } = require('typescript');
 
-function readFile(path) {
+const files = ['hook-ipc', 'start-native'];
+
+function readString(path) {
   return readFileSync(join(__dirname, path), { encoding: 'utf8' });
 }
 
-const ipcModule = readFile('../src/ipc.ts');
+const ipcModule = readString('../src/ipc.ts');
 
 function transpileFile(path) {
-  let file = readFile(path)
-    // Since we're not doing proper bundling we need to replace the ipc import with the code 😋!
-    .replace("import { IPC } from '../ipc';", ipcModule);
+  // Since we're not doing proper bundling we need to replace the ipc import with the code 😋
+  let file = readString(path).replace("import { IPC } from '../ipc';", ipcModule);
 
-  return ts.transpile(file, { removeComments: true });
+  return transpile(file, { removeComments: true });
 }
 
-const template = readFile('../src/preload/bundled.template.ts');
+const template = readString('../src/preload/bundled.template.ts');
 
-writeFileSync(
-  join(__dirname, '../src/preload/bundled.ts'),
-  template
-    .replace('{{hook-ipc.js}}', transpileFile('../src/preload/hook-ipc.ts'))
-    .replace('{{start-native.js}}', transpileFile('../src/preload/start-native.ts')),
+// Replace all the keys in the template with transpiled code
+const code = files.reduce(
+  (tpl, file) => tpl.replace(`{{${file}}}`, transpileFile(`../src/preload/${file}.ts`)),
+  template,
 );
+
+writeFileSync(join(__dirname, '../src/preload/bundled.ts'), code);
